requirements.txt
transformers
torch
gradio
PyMuPDF
openai
langchain-openai
langchain-core
python-dotenv
